refactor: adds commons folder
