+++ src/common/resolvePaths.js
