+++ src/common/stream.js
